requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
# Pin passlib & bcrypt to compatible versions to avoid AttributeError: module 'bcrypt' has no attribute '__about__'
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
